requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.9
SQLAlchemy==2.0.20
pytest==8.2.2
pytest-cov==4.1.0
flake8==6.1.0
pandas==2.1.1
torch==2.1.0
Flask-Cors>=3.0.10
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0
torchvision>=0.18.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
Pillow>=9.0.0
bcrypt==4.1.2
opencv-python==4.8.1.78
numpy
keras-facenet==0.3.2
tensorflow>=2.16.0
kagglehub
scikit-learn
